Grid: Handle single length vent size
No examples in example or puzzle input may just be by chance, so I want to ensure the code handles the situation.

This requires special handling since I use streams.

If x start and end location is identical, an infinite stream is handled so we can zip the y stream to have a vertical vent.

Similar situation for horizontal vents.

A single length vent is both horizontal and vertical so both x and y stream would be infinite thus the zipped stream would be infinite unless it is explicitly limited to length one.