LanternFishCounts: Simplify further - no need to move values in the array
Fish giving birth on a given day will next give birth in DAYS_BETWEEN_BIRTHS time, so we can add the count of fish giving birth today to the location we'll read from on that day.

After a full cycle around the count array, the fish born today will give birth, so we keep the count at the current location as is.